+++ CHANGELOG.md
@@ -11,6 +11,8 @@ BUG FIXES:
 
 - fix(dbaas): improve terraform refresh #488
 - dbaas: handle correctly terraform refresh for db and user #490
+- fix(iam): improve terraform refresh #494
+- fix(iam): prevent type error when rules are empty for resource exoscale_iam_org_policy
 
 BREAKING CHANGES:
 - update dbaas schema, check the [migration guide](docs/guides/migration-of-dbaas-from-v0_67_x-to-v0_68_x.md) #488